hocs 폴더 생성해서 withHeader HOC를 이동시킴
