+++ src/pages/chat-list.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import withHeader from 'components/layouts/header'
+import withHeader from 'hocs/withHeader'
 
 const ChatListPage = () => <h1>chatlist page</h1>
