+++ src/pages/community.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import withHeader from 'components/layouts/header'
+import withHeader from 'hocs/withHeader'
 
 const CommunityPage = () => <h1>community page</h1>
