+++ src/pages/my-info.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import withHeader from 'components/layouts/header'
+import withHeader from 'hocs/withHeader'
 
 const MyInfoPage = () => <h1>myinfo page</h1>
