+++ src/pages/search.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import withHeader from 'components/layouts/header'
+import withHeader from 'hocs/withHeader'
 
 const SearchPage = () => <h1>search page</h1>
